fix(Healthcheck): separate issue location, increase gaps
