+++ src/containers/Tenant/Healthcheck/Healthcheck.scss
@@ -53,7 +53,7 @@
         }
     }
     &__issue-content {
-        --g-definition-list-item-gap: var(--g-spacing-2);
+        --g-definition-list-item-gap: var(--g-spacing-3);
         width: 100%;
     }
     &__issue-divider {
